feat: 测试页面迁移进packages + 根目录启动示例
