+++ package.json
@@ -3,6 +3,7 @@
   "private": true,
   "description": "前端监控SDK，可用来收集并上报：代码报错、性能数据、用户行为、加载资源、个性化指标等数据",
   "scripts": {
+    "start": "cd packages/example && pnpm run start",
     "build": "rollup -c",
     "prepare": "husky install",
     "commit": "git-cz",
