+++ packages/example/babel.config.js
