+++ packages/example/public/favicon.ico
