+++ packages/example/src/App.vue
