+++ packages/example/src/assets/logo.png
