+++ packages/example/src/main.js
@@ -5,13 +5,9 @@ import store from './store';
 import ElementUI from 'element-ui';
 import 'element-ui/lib/theme-chalk/index.css';
 
-import webSee from '../../packages/core/src';
-import performance from '../../packages/performance/src';
-import recordscreen from '../../packages/recordscreen/src';
-
-// import webSee from '@websee/core';
-// import performance from '@websee/performance';
-// import recordscreen from '@websee/recordscreen';
+import webSee from '@websee/core';
+import performance from '@websee/performance';
+import recordscreen from '@websee/recordscreen';
 
 Vue.use(webSee, {
   dsn: 'http://localhost:8080/reportData',
