+++ packages/example/src/store/index.js
