+++ packages/example/src/utils/recordScreen.js
