+++ packages/example/vue.config.js
